requirements.txt
Django==1.11.4
django-widget-tweaks==1.4.1
django-admin-view-permission
django-environ
feedparser
pycrypto
python-dotenv

